requirements.txt
music21>=9.1
chromadb>=1.0
ollama>=0.4
python-dotenv>=1.0
requests>=2.31
mido>=1.3
pretty_midi>=0.2.10
fastapi>=0.115
uvicorn[standard]>=0.34
python-multipart>=0.0.18
sse-starlette>=2.2
httpx>=0.28
pytest-asyncio>=0.25
pytest-cov>=6.0
